refactor(test): extract usecase layer tests to tests/unit/
Move 43 public API tests from 8 usecase source files to tests/unit/usecase/.
Delete 52 private API tests that accessed non-public methods or constructed
#[non_exhaustive] types (Tag, QueryParams, ShamirShare, CFragData, etc.).

Moved tests: dto(4), error(14), core/crypto(9), workflow/container(3),
workflow/secret_sharing_service(8), workflow/secret_recovery_service(5).

Deleted tests: core/crypto(4), core/storage(3), core/contract_storage(3),
workflow/secret_sharing_service(16), workflow/secret_recovery_service(26).

Test counts: 186 inline + 159 external + 60 integration = 405 total.
