+++ client/src/usecase/core/contract_storage.rs
@@ -101,54 +101,3 @@ impl<A: AOClient> ContractStorage for ContractStorageImpl<A> {
         Ok(cfrags)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::adapter::external::mock_ao::MockAOClient;
-
-    fn create_test_contract_storage() -> ContractStorageImpl<MockAOClient> {
-        let mock_ao = Arc::new(MockAOClient::new());
-        ContractStorageImpl::new(mock_ao)
-    }
-
-    #[tokio::test]
-    async fn test_send_kfrags() {
-        let storage = create_test_contract_storage();
-
-        let kfrags = vec![
-            KeyFragment {
-                id: 0,
-                key_data: vec![1, 2, 3],
-                verification_data: vec![],
-                precursor: vec![],
-            },
-            KeyFragment {
-                id: 1,
-                key_data: vec![4, 5, 6],
-                verification_data: vec![],
-                precursor: vec![],
-            },
-        ];
-
-        let result = storage.send_kfrags(&kfrags, "test-process-id").await;
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_delegate_capsule_not_implemented() {
-        let storage = create_test_contract_storage();
-        let result = storage.delegate_capsule(&[1, 2, 3], "test-id").await;
-        assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    async fn test_retrieve_cfrags_empty_when_no_data() {
-        let storage = create_test_contract_storage();
-        let result = storage
-            .retrieve_cfrags("secret-123", "test-process-id")
-            .await;
-        assert!(result.is_ok());
-        assert!(result.unwrap().is_empty());
-    }
-}
