+++ client/src/usecase/core/storage.rs
@@ -358,138 +358,3 @@ impl Default for ArweaveStorageServiceImpl {
         Self::new(StorageConfig::default())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_store_and_retrieve() {
-        println!("\n=== ArweaveStorageService: Store and Retrieve Test ===");
-        println!("【テスト内容】: Arweaveへのデータ保存と取得機能を検証");
-        println!("【テスト対象】: store_data()とretrieve_data()メソッド");
-        println!("【期待結果】: データとタグが正しく保存・取得される");
-
-        let service = ArweaveStorageServiceImpl::default();
-
-        // テストデータの準備
-        let data = b"test data";
-        let data_hex: Vec<String> = data.iter().map(|b| format!("{b:02x}")).collect();
-        let data_text = std::str::from_utf8(data).unwrap();
-        let data_len = data.len();
-        println!("\n1. 保存するデータを準備:");
-        println!("   - テキスト: \"{data_text}\"");
-        println!("   - バイナリ: [{}]", data_hex.join(" "));
-        println!("   - サイズ: {data_len} bytes");
-
-        let tags = vec![
-            Tag {
-                name: "Type".to_string(),
-                value: "Test".to_string(),
-            },
-            Tag {
-                name: "Version".to_string(),
-                value: "1.0".to_string(),
-            },
-        ];
-        println!("\n2. タグを設定:");
-        for tag in &tags {
-            println!("   - {}: {}", tag.name, tag.value);
-        }
-
-        // データの保存
-        println!("\n3. Arweaveにデータを保存...");
-        let tx_id = service.store_data(data, tags).unwrap();
-        println!("   ✓ 保存成功!");
-        println!("   トランザクションID: {tx_id}");
-        assert!(!tx_id.is_empty());
-
-        // データの取得
-        println!("\n4. 保存したデータを取得...");
-        let retrieved = service.retrieve_data(&tx_id).unwrap();
-        println!("   ✓ 取得成功!");
-
-        // 取得したデータの検証
-        println!("\n5. 取得したデータを検証:");
-        println!("   - データサイズ: {} bytes", retrieved.data.len());
-        println!(
-            "   - データ内容: \"{}\"",
-            std::str::from_utf8(&retrieved.data).unwrap()
-        );
-        println!("   - タグ数: {}", retrieved.tags.len());
-
-        assert_eq!(retrieved.data, data);
-        println!("   ✓ データが一致");
-
-        assert_eq!(retrieved.tags.len(), 2);
-        println!("   ✓ タグ数が一致");
-
-        // タグの詳細を表示
-        println!("\n6. 取得したタグの詳細:");
-        for tag in &retrieved.tags {
-            println!("   - {}: {}", tag.name, tag.value);
-        }
-
-        println!("\n✅ テスト成功: データの保存・取得が正常に動作しました！");
-    }
-
-    #[test]
-    fn test_query_by_tags() {
-        let service = ArweaveStorageServiceImpl::default();
-
-        // 複数のトランザクションを作成
-        for i in 0..3 {
-            let data = format!("data {i}").into_bytes();
-            let tags = vec![
-                Tag {
-                    name: "Type".to_string(),
-                    value: "Test".to_string(),
-                },
-                Tag {
-                    name: "Index".to_string(),
-                    value: i.to_string(),
-                },
-            ];
-            service.store_data(&data, tags).unwrap();
-        }
-
-        // タグでクエリ
-        let params = QueryParams {
-            tags: vec![Tag {
-                name: "Type".to_string(),
-                value: "Test".to_string(),
-            }],
-            limit: Some(2),
-            sort_by: Some(SortBy::Timestamp(SortOrder::Descending)),
-        };
-
-        let results = service.query_by_tags(params).unwrap();
-        assert_eq!(results.len(), 2);
-    }
-
-    #[test]
-    fn test_batch_store() {
-        let service = ArweaveStorageServiceImpl::default();
-
-        let items = vec![
-            (
-                b"data1".to_vec(),
-                vec![Tag {
-                    name: "ID".to_string(),
-                    value: "1".to_string(),
-                }],
-            ),
-            (
-                b"data2".to_vec(),
-                vec![Tag {
-                    name: "ID".to_string(),
-                    value: "2".to_string(),
-                }],
-            ),
-        ];
-
-        let result = service.batch_store(items).unwrap();
-        assert_eq!(result.successful.len(), 2);
-        assert_eq!(result.failed.len(), 0);
-    }
-}
