+++ client/src/usecase/workflow/secret_sharing_service.rs
@@ -362,893 +362,3 @@ impl<C: CryptoService, ST: StorageService> SecretSharingWorkflowService
         Ok(SecretStatus::Created)
     }
 }
-
-#[cfg(test)]
-#[allow(
-    dead_code,
-    clippy::uninlined_format_args,
-    clippy::indexing_slicing,
-    clippy::redundant_clone,
-    clippy::cast_possible_truncation,
-    clippy::arithmetic_side_effects,
-    clippy::unwrap_used,
-    clippy::type_complexity
-)]
-mod tests {
-    use super::*;
-    use crate::adapter::external::mock_ao::MockAOClient;
-    use crate::service::error::ServiceResult;
-    use crate::usecase::core::contract_storage::ContractStorageImpl;
-    use crate::usecase::core::crypto::{
-        CryptoService as CoreCryptoService, CryptoServiceImpl as CoreCryptoServiceImpl,
-    };
-    use crate::usecase::core::storage::{
-        ArweaveStorageServiceImpl, ArweaveTransaction, BatchResult, QueryParams, Tag,
-        TransactionStatus,
-    };
-    use crate::usecase::service::{
-        CryptoServiceImpl as ServiceCryptoServiceImpl,
-        StorageServiceImpl as ServiceStorageServiceImpl,
-    };
-    use std::sync::RwLock;
-
-    type TestCryptoService = ServiceCryptoServiceImpl<CoreCryptoServiceImpl>;
-    type TestStorageService =
-        ServiceStorageServiceImpl<ArweaveStorageServiceImpl, ContractStorageImpl<MockAOClient>>;
-
-    // ========================================================================
-    // MockStorageService (Task 8)
-    // ========================================================================
-
-    /// Mock implementation of StorageService for testing
-    struct MockStorageService {
-        /// Stores all data passed to store_data()
-        stored_data: Arc<RwLock<Vec<StoredItem>>>,
-        /// Stores all batch items passed to batch_store()
-        batch_items: Arc<RwLock<Vec<(Vec<u8>, Vec<Tag>)>>>,
-        /// Stores kFrag calls
-        kfrag_calls: Arc<RwLock<Vec<KFragCall>>>,
-        /// Stores query calls
-        query_results: Arc<RwLock<Vec<ArweaveTransaction>>>,
-        /// Transaction counter
-        tx_counter: Arc<RwLock<u64>>,
-        /// Controls whether send_kfrag should fail
-        should_fail_kfrag: Arc<RwLock<bool>>,
-        /// Controls whether store_data should fail
-        should_fail_store: Arc<RwLock<bool>>,
-        /// Controls whether batch_store should partially fail (odd indices fail)
-        should_fail_batch_partial: Arc<RwLock<bool>>,
-    }
-
-    #[derive(Debug, Clone)]
-    struct StoredItem {
-        data: Vec<u8>,
-        tags: Vec<Tag>,
-        tx_id: String,
-    }
-
-    #[derive(Debug, Clone)]
-    struct KFragCall {
-        kfrag_count: usize,
-        owner_process_id: String,
-    }
-
-    impl MockStorageService {
-        fn new() -> Self {
-            Self {
-                stored_data: Arc::new(RwLock::new(Vec::new())),
-                batch_items: Arc::new(RwLock::new(Vec::new())),
-                kfrag_calls: Arc::new(RwLock::new(Vec::new())),
-                query_results: Arc::new(RwLock::new(Vec::new())),
-                tx_counter: Arc::new(RwLock::new(0)),
-                should_fail_kfrag: Arc::new(RwLock::new(false)),
-                should_fail_store: Arc::new(RwLock::new(false)),
-                should_fail_batch_partial: Arc::new(RwLock::new(false)),
-            }
-        }
-
-        fn get_stored_data(&self) -> Vec<StoredItem> {
-            self.stored_data.read().unwrap().clone()
-        }
-
-        fn get_batch_items(&self) -> Vec<(Vec<u8>, Vec<Tag>)> {
-            self.batch_items.read().unwrap().clone()
-        }
-
-        fn get_kfrag_calls(&self) -> Vec<KFragCall> {
-            self.kfrag_calls.read().unwrap().clone()
-        }
-
-        fn set_query_results(&self, results: Vec<ArweaveTransaction>) {
-            *self.query_results.write().unwrap() = results;
-        }
-
-        fn set_should_fail_kfrag(&self, should_fail: bool) {
-            *self.should_fail_kfrag.write().unwrap() = should_fail;
-        }
-
-        fn set_should_fail_store(&self, should_fail: bool) {
-            *self.should_fail_store.write().unwrap() = should_fail;
-        }
-
-        fn set_should_fail_batch_partial(&self, should_fail: bool) {
-            *self.should_fail_batch_partial.write().unwrap() = should_fail;
-        }
-
-        fn build_partial_failure_result(&self, item_count: usize) -> BatchResult {
-            let mut successful = Vec::new();
-            let mut failed = Vec::new();
-            for index in 0..item_count {
-                if index % 2 == 1 {
-                    failed.push((index.to_string(), "Mock partial failure".to_string()));
-                } else {
-                    successful.push(self.generate_tx_id());
-                }
-            }
-            BatchResult { successful, failed }
-        }
-
-        fn generate_tx_id(&self) -> String {
-            let mut counter = self.tx_counter.write().unwrap();
-            *counter += 1;
-            format!("mock_tx_{:016x}", *counter)
-        }
-    }
-
-    #[async_trait]
-    impl StorageService for MockStorageService {
-        fn store_data(&self, data: &[u8], tags: Vec<Tag>) -> ServiceResult<String> {
-            if *self.should_fail_store.read().unwrap() {
-                return Err(crate::service::error::ServiceError::storage_error(
-                    "Mock storage failure",
-                ));
-            }
-
-            let tx_id = self.generate_tx_id();
-            self.stored_data.write().unwrap().push(StoredItem {
-                data: data.to_vec(),
-                tags,
-                tx_id: tx_id.clone(),
-            });
-            Ok(tx_id)
-        }
-
-        fn retrieve_data(&self, transaction_id: &str) -> ServiceResult<ArweaveTransaction> {
-            let stored = self.stored_data.read().unwrap();
-            stored
-                .iter()
-                .find(|item| item.tx_id == transaction_id)
-                .map(|item| ArweaveTransaction {
-                    id: item.tx_id.clone(),
-                    data: item.data.clone(),
-                    tags: item.tags.clone(),
-                    timestamp: 0,
-                })
-                .ok_or_else(|| {
-                    crate::service::error::ServiceError::not_found(format!(
-                        "Transaction {} not found",
-                        transaction_id
-                    ))
-                })
-        }
-
-        fn query_by_tags(&self, _params: QueryParams) -> ServiceResult<Vec<ArweaveTransaction>> {
-            Ok(self.query_results.read().unwrap().clone())
-        }
-
-        fn check_transaction_status(
-            &self,
-            _transaction_id: &str,
-        ) -> ServiceResult<TransactionStatus> {
-            Ok(TransactionStatus::Confirmed)
-        }
-
-        fn batch_store(&self, items: Vec<(Vec<u8>, Vec<Tag>)>) -> ServiceResult<BatchResult> {
-            if *self.should_fail_store.read().unwrap() {
-                return Err(crate::service::error::ServiceError::storage_error(
-                    "Mock batch storage failure",
-                ));
-            }
-
-            self.batch_items.write().unwrap().extend(items.clone());
-
-            if *self.should_fail_batch_partial.read().unwrap() {
-                return Ok(self.build_partial_failure_result(items.len()));
-            }
-
-            let successful: Vec<String> = items.iter().map(|_| self.generate_tx_id()).collect();
-            Ok(BatchResult {
-                successful,
-                failed: Vec::new(),
-            })
-        }
-
-        fn exists(&self, transaction_id: &str) -> ServiceResult<bool> {
-            let stored = self.stored_data.read().unwrap();
-            Ok(stored.iter().any(|item| item.tx_id == transaction_id))
-        }
-
-        fn update_tags(&self, transaction_id: &str, new_tags: Vec<Tag>) -> ServiceResult<String> {
-            let data = self.retrieve_data(transaction_id)?.data;
-            self.store_data(&data, new_tags)
-        }
-
-        async fn send_kfrags_to_contract(
-            &self,
-            kfrags: &[KeyFragment],
-            contract_id: &str,
-        ) -> ServiceResult<()> {
-            if *self.should_fail_kfrag.read().unwrap() {
-                return Err(crate::service::error::ServiceError::ao_network_error(
-                    "Mock kFrag send failure",
-                ));
-            }
-
-            self.kfrag_calls.write().unwrap().push(KFragCall {
-                kfrag_count: kfrags.len(),
-                owner_process_id: contract_id.to_string(),
-            });
-            Ok(())
-        }
-
-        async fn delegate_capsule(
-            &self,
-            _capsule_data: &[u8],
-            _contract_id: &str,
-        ) -> ServiceResult<()> {
-            Err(crate::service::error::ServiceError::System(
-                crate::service::error::SystemException::Internal("Not implemented".to_string()),
-            ))
-        }
-
-        async fn retrieve_cfrags(
-            &self,
-            _secret_id: &str,
-            _requester_process_id: &str,
-        ) -> ServiceResult<Vec<crate::usecase::core::crypto::CFragData>> {
-            Err(crate::service::error::ServiceError::System(
-                crate::service::error::SystemException::Internal("Not implemented".to_string()),
-            ))
-        }
-
-        fn retrieve_encrypted_shares(&self, _secret_id: &str) -> ServiceResult<Vec<Vec<u8>>> {
-            Err(crate::service::error::ServiceError::System(
-                crate::service::error::SystemException::Internal("Not implemented".to_string()),
-            ))
-        }
-    }
-
-    // ========================================================================
-    // Test Helper Functions
-    // ========================================================================
-
-    fn create_test_service()
-    -> SecretSharingWorkflowServiceImpl<TestCryptoService, TestStorageService> {
-        let core_crypto = Arc::new(CoreCryptoServiceImpl::new());
-        let crypto = Arc::new(ServiceCryptoServiceImpl::new(Arc::clone(&core_crypto)));
-        let mock_ao = Arc::new(MockAOClient::new());
-        let arweave = Arc::new(ArweaveStorageServiceImpl::default());
-        let contract = Arc::new(ContractStorageImpl::new(mock_ao));
-        let storage = Arc::new(ServiceStorageServiceImpl::new(arweave, contract));
-        SecretSharingWorkflowServiceImpl::new(crypto, storage)
-    }
-
-    fn create_mock_service() -> (
-        SecretSharingWorkflowServiceImpl<TestCryptoService, MockStorageService>,
-        Arc<MockStorageService>,
-    ) {
-        let core_crypto = Arc::new(CoreCryptoServiceImpl::new());
-        let crypto = Arc::new(ServiceCryptoServiceImpl::new(core_crypto));
-        let storage = Arc::new(MockStorageService::new());
-        let service = SecretSharingWorkflowServiceImpl::new(crypto, storage.clone());
-        (service, storage)
-    }
-
-    fn create_test_request(crypto: &CoreCryptoServiceImpl) -> SecretSharingRequest {
-        let (owner_sk, owner_pk) = crypto.generate_keypair().unwrap();
-        let (_requester_sk, requester_pk) = crypto.generate_keypair().unwrap();
-
-        SecretSharingRequest {
-            secret: b"Test secret data for sharing".to_vec(),
-            owner_secret_key: owner_sk,
-            owner_public_key: owner_pk,
-            requester_public_key: requester_pk,
-            threshold: 3,
-            total_shares: 5,
-            owner_process_id: "owner-process-123".to_string(),
-            metadata: None,
-        }
-    }
-
-    // ========================================================================
-    // Validation Tests (Task 14)
-    // ========================================================================
-
-    #[test]
-    fn test_validate_request_valid() {
-        println!("\n=== test_validate_request_valid ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let request = create_test_request(&crypto);
-        println!(
-            "  Created request: threshold={}, total_shares={}, secret_len={}",
-            request.threshold,
-            request.total_shares,
-            request.secret.len()
-        );
-
-        let result = service.validate_request(&request);
-        println!("  Validation result: {:?}", result.is_ok());
-        assert!(result.is_ok());
-        println!("  [PASS] Valid request accepted");
-    }
-
-    #[test]
-    fn test_phase1_invalid_threshold_below_min() {
-        println!("\n=== test_phase1_invalid_threshold_below_min ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let mut request = create_test_request(&crypto);
-        request.threshold = 1; // Below MIN_THRESHOLD
-        println!("  Testing threshold=1 (below MIN_THRESHOLD=2)");
-
-        let result = service.validate_request(&request);
-        println!("  Result: {:?}", result);
-        assert!(matches!(result, Err(WorkflowError::ValidationError(_))));
-
-        if let Err(WorkflowError::ValidationError(msg)) = result {
-            assert!(msg.contains("at least"));
-            println!("  [PASS] Correctly rejected with message: {}", msg);
-        }
-    }
-
-    #[test]
-    fn test_phase1_invalid_threshold_exceeds_total() {
-        println!("\n=== test_phase1_invalid_threshold_exceeds_total ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let mut request = create_test_request(&crypto);
-        request.threshold = 6;
-        request.total_shares = 5;
-        println!("  Testing threshold=6 > total_shares=5");
-
-        let result = service.validate_request(&request);
-        println!("  Result: {:?}", result);
-        assert!(matches!(result, Err(WorkflowError::ValidationError(_))));
-
-        if let Err(WorkflowError::ValidationError(msg)) = result {
-            assert!(msg.contains("cannot exceed"));
-            println!("  [PASS] Correctly rejected: {}", msg);
-        }
-    }
-
-    #[test]
-    fn test_phase1_invalid_total_shares_exceeds_max() {
-        println!("\n=== test_phase1_invalid_total_shares_exceeds_max ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let mut request = create_test_request(&crypto);
-        request.total_shares = 255; // Exceeds MAX_SHARES
-        request.threshold = 3;
-        println!("  Testing total_shares=255 (exceeds MAX)");
-
-        let result = service.validate_request(&request);
-        println!("  Result: {:?}", result);
-        assert!(matches!(result, Err(WorkflowError::ValidationError(_))));
-
-        if let Err(WorkflowError::ValidationError(msg)) = result {
-            assert!(msg.contains("cannot exceed") || msg.contains("255"));
-            println!("  [PASS] Correctly rejected: {}", msg);
-        }
-    }
-
-    #[test]
-    fn test_phase1_empty_secret() {
-        println!("\n=== test_phase1_empty_secret ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let mut request = create_test_request(&crypto);
-        request.secret = vec![];
-        println!("  Testing empty secret");
-
-        let result = service.validate_request(&request);
-        println!("  Result: {:?}", result);
-        assert!(matches!(result, Err(WorkflowError::ValidationError(_))));
-
-        if let Err(WorkflowError::ValidationError(msg)) = result {
-            assert!(msg.contains("empty"));
-            println!("  [PASS] Correctly rejected: {}", msg);
-        }
-    }
-
-    #[test]
-    fn test_phase1_invalid_owner_public_key() {
-        println!("\n=== test_phase1_invalid_owner_public_key ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let mut request = create_test_request(&crypto);
-        request.owner_public_key = PublicKey { key_data: vec![] };
-        println!("  Testing empty owner public key");
-
-        let result = service.validate_request(&request);
-        println!("  Result: {:?}", result);
-        assert!(matches!(result, Err(WorkflowError::ValidationError(_))));
-
-        if let Err(WorkflowError::ValidationError(msg)) = result {
-            assert!(msg.contains("Owner public key"));
-            println!("  [PASS] Correctly rejected: {}", msg);
-        }
-    }
-
-    #[test]
-    fn test_phase1_invalid_requester_public_key() {
-        println!("\n=== test_phase1_invalid_requester_public_key ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let mut request = create_test_request(&crypto);
-        request.requester_public_key = PublicKey { key_data: vec![] };
-        println!("  Testing empty requester public key");
-
-        let result = service.validate_request(&request);
-        println!("  Result: {:?}", result);
-        assert!(matches!(result, Err(WorkflowError::ValidationError(_))));
-
-        if let Err(WorkflowError::ValidationError(msg)) = result {
-            assert!(msg.contains("Requester public key"));
-            println!("  [PASS] Correctly rejected: {}", msg);
-        }
-    }
-
-    #[test]
-    fn test_phase1_empty_owner_process_id() {
-        println!("\n=== test_phase1_empty_owner_process_id ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let mut request = create_test_request(&crypto);
-        request.owner_process_id = String::new();
-        println!("  Testing empty owner process ID");
-
-        let result = service.validate_request(&request);
-        println!("  Result: {:?}", result);
-        assert!(matches!(result, Err(WorkflowError::ValidationError(_))));
-
-        if let Err(WorkflowError::ValidationError(msg)) = result {
-            assert!(msg.contains("process ID"));
-            println!("  [PASS] Correctly rejected: {}", msg);
-        }
-    }
-
-    // ========================================================================
-    // Workflow Tests (Task 15)
-    // ========================================================================
-
-    #[test]
-    fn test_phase1_generates_symmetric_key() {
-        println!("\n=== test_phase1_generates_symmetric_key ===");
-        let service = create_test_service();
-
-        println!("  Generating symmetric key...");
-        let result = service.generate_symmetric_key();
-        assert!(result.is_ok());
-
-        let key = result.unwrap();
-        println!("  Key length: {} bytes", key.len());
-        assert_eq!(key.len(), 32); // AES-256 key
-        println!("  [PASS] Generated valid AES-256 key (32 bytes)");
-    }
-
-    #[test]
-    fn test_phase1_generates_symmetric_key_randomness() {
-        println!("\n=== test_phase1_generates_symmetric_key_randomness ===");
-        let service = create_test_service();
-
-        println!("  Generating two keys to verify randomness...");
-        let key1 = service.generate_symmetric_key().unwrap();
-        let key2 = service.generate_symmetric_key().unwrap();
-
-        println!("  Key1: {:02x?}...", &key1[..8]);
-        println!("  Key2: {:02x?}...", &key2[..8]);
-        // Keys should be different (random)
-        assert_ne!(key1, key2);
-        println!("  [PASS] Keys are different (cryptographically random)");
-    }
-
-    #[test]
-    fn test_phase1_splits_secret_shamir() {
-        println!("\n=== test_phase1_splits_secret_shamir ===");
-        let service = create_test_service();
-        let secret = b"Test secret data";
-        println!(
-            "  Secret: {:?} ({} bytes)",
-            String::from_utf8_lossy(secret),
-            secret.len()
-        );
-
-        println!("  Splitting with threshold=3, total_shares=5...");
-        let result = service.split_secret(secret, 3, 5);
-        assert!(result.is_ok());
-
-        let shares = result.unwrap();
-        println!("  Generated {} shares", shares.len());
-        assert_eq!(shares.len(), 5);
-
-        // Each share should have a unique index
-        let indices: Vec<u8> = shares.iter().map(|s| s.index).collect();
-        println!("  Share indices: {:?}", indices);
-        let mut unique_indices = indices.clone();
-        unique_indices.sort_unstable();
-        unique_indices.dedup();
-        assert_eq!(indices.len(), unique_indices.len());
-        println!("  [PASS] All shares have unique indices");
-    }
-
-    #[test]
-    fn test_phase1_encrypts_shares_aes_gcm() {
-        println!("\n=== test_phase1_encrypts_shares_aes_gcm ===");
-        let service = create_test_service();
-        let symmetric_key = service.generate_symmetric_key().unwrap();
-        println!("  Generated symmetric key");
-
-        let shares = vec![
-            ShamirShare {
-                index: 1,
-                share_data: b"share data 1".to_vec(),
-            },
-            ShamirShare {
-                index: 2,
-                share_data: b"share data 2".to_vec(),
-            },
-        ];
-        println!("  Encrypting {} shares with AES-GCM...", shares.len());
-
-        let result = service.encrypt_shares(&shares, &symmetric_key);
-        assert!(result.is_ok());
-
-        let encrypted = result.unwrap();
-        assert_eq!(encrypted.len(), 2);
-
-        // Encrypted data should be larger than original (includes nonce + tag)
-        for (i, enc) in encrypted.iter().enumerate() {
-            println!(
-                "  Share {}: {} bytes -> {} bytes (encrypted)",
-                i,
-                shares[i].share_data.len(),
-                enc.len()
-            );
-            assert!(enc.len() > shares[i].share_data.len());
-        }
-        println!("  [PASS] All shares encrypted successfully");
-    }
-
-    #[test]
-    fn test_phase1_creates_capsule_for_symmetric_key() {
-        println!("\n=== test_phase1_creates_capsule_for_symmetric_key ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let (_owner_sk, owner_pk) = crypto.generate_keypair().unwrap();
-        let symmetric_key = service.generate_symmetric_key().unwrap();
-        println!("  Generated owner keypair and symmetric key");
-
-        println!("  Creating PRE capsule...");
-        let result = service.create_capsule(&owner_pk, &symmetric_key);
-        assert!(result.is_ok());
-
-        let (capsule, ciphertext) = result.unwrap();
-        println!("  Capsule size: {} bytes", capsule.capsule_bytes.len());
-        println!("  Ciphertext size: {} bytes", ciphertext.len());
-        assert!(!capsule.capsule_bytes.is_empty());
-        assert!(!ciphertext.is_empty());
-        println!("  [PASS] PRE capsule created successfully");
-    }
-
-    #[test]
-    fn test_phase1_generates_kfrags() {
-        println!("\n=== test_phase1_generates_kfrags ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let request = create_test_request(&crypto);
-        println!(
-            "  threshold={}, total_shares={}",
-            request.threshold, request.total_shares
-        );
-
-        println!("  Generating kFrags...");
-        let result = service.generate_kfrags(&request);
-        assert!(result.is_ok());
-
-        let kfrags = result.unwrap();
-        println!("  Generated {} kFrags", kfrags.len());
-        assert_eq!(kfrags.len(), 5); // total_shares
-
-        // Each kFrag should have data
-        for (i, kfrag) in kfrags.iter().enumerate() {
-            println!("  kFrag[{}]: {} bytes", i, kfrag.key_data.len());
-            assert!(!kfrag.key_data.is_empty());
-        }
-        println!("  [PASS] All kFrags generated with valid data");
-    }
-
-    #[test]
-    fn test_phase1_kfrags_count_matches_n() {
-        println!("\n=== test_phase1_kfrags_count_matches_n ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-
-        // Test with different n values
-        for n in [3, 5, 7, 10] {
-            let mut request = create_test_request(&crypto);
-            request.total_shares = n;
-            request.threshold = 2;
-
-            println!("  Testing n={}: generating kFrags...", n);
-            let kfrags = service.generate_kfrags(&request).unwrap();
-            println!("  Generated {} kFrags for n={}", kfrags.len(), n);
-            assert_eq!(
-                kfrags.len() as u8,
-                n,
-                "kFrag count should match total_shares"
-            );
-        }
-        println!("  [PASS] kFrag count matches n for all test values");
-    }
-
-    #[tokio::test]
-    async fn test_phase1_returns_complete_result() {
-        println!("\n=== test_phase1_returns_complete_result ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let request = create_test_request(&crypto);
-        let owner_pk = request.owner_public_key.clone();
-        println!("  Executing complete PHASE 1 workflow...");
-
-        let result = service.execute_secret_sharing(request).await;
-        assert!(result.is_ok());
-
-        let result = result.unwrap();
-        println!("  Result:");
-        println!("    secret_id: {}", result.secret_id);
-        println!("    capsule_tx_id: {}", result.capsule_tx_id);
-        println!("    share_tx_ids: {} items", result.share_tx_ids.len());
-        println!("    kfrag_count: {}", result.kfrag_count);
-
-        // Verify all fields are populated
-        assert!(!result.secret_id.as_str().is_empty());
-        assert!(!result.capsule_tx_id.is_empty());
-        assert_eq!(result.share_tx_ids.len(), 5);
-        assert_eq!(result.kfrag_count, 5);
-        assert_eq!(result.owner_public_key.key_data, owner_pk.key_data);
-        println!("  [PASS] Complete result returned with all fields populated");
-    }
-
-    #[tokio::test]
-    async fn test_phase1_complete_flow() {
-        println!("\n=== test_phase1_complete_flow ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let original_secret = b"This is a very important secret message!".to_vec();
-        println!(
-            "  Original secret: {:?}",
-            String::from_utf8_lossy(&original_secret)
-        );
-
-        let mut request = create_test_request(&crypto);
-        request.secret = original_secret.clone();
-        request.threshold = 3;
-        request.total_shares = 5;
-        println!("  Parameters: threshold=3, total_shares=5");
-
-        println!("  Executing PHASE 1 workflow...");
-        let result = service.execute_secret_sharing(request).await;
-        assert!(result.is_ok());
-
-        let result = result.unwrap();
-        println!("  Result:");
-        println!("    kfrag_count: {}", result.kfrag_count);
-        println!("    share_tx_ids: {:?}", result.share_tx_ids);
-        println!("    capsule_tx_id: {}", result.capsule_tx_id);
-        assert_eq!(result.kfrag_count, 5);
-        assert_eq!(result.share_tx_ids.len(), 5);
-
-        // Verify transaction IDs have correct format
-        for tx_id in &result.share_tx_ids {
-            assert!(tx_id.starts_with("tx_"));
-        }
-        assert!(result.capsule_tx_id.starts_with("tx_"));
-        println!("  [PASS] Complete PHASE 1 flow executed successfully");
-    }
-
-    #[tokio::test]
-    async fn test_phase1_minimum_threshold() {
-        println!("\n=== test_phase1_minimum_threshold ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let mut request = create_test_request(&crypto);
-        request.threshold = constants::MIN_THRESHOLD;
-        request.total_shares = constants::MIN_THRESHOLD;
-        println!(
-            "  Testing minimum threshold: k=n={}",
-            constants::MIN_THRESHOLD
-        );
-
-        let result = service.execute_secret_sharing(request).await;
-        println!("  Result: {:?}", result.is_ok());
-        assert!(result.is_ok());
-        println!("  [PASS] Minimum threshold accepted");
-    }
-
-    #[tokio::test]
-    async fn test_phase1_max_secret_size() {
-        println!("\n=== test_phase1_max_secret_size ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let mut request = create_test_request(&crypto);
-
-        // Test with max supported secret size (63 bytes - DATA_SIZE is 64, 1 byte for length prefix)
-        request.secret = vec![0xAB; 63];
-        println!("  Testing max secret size: {} bytes", request.secret.len());
-
-        let result = service.execute_secret_sharing(request).await;
-        println!("  Result: {:?}", result.is_ok());
-        assert!(result.is_ok());
-        println!("  [PASS] Max secret size (63 bytes) accepted");
-    }
-
-    // ========================================================================
-    // Error Handling & Status Tests (Task 16)
-    // ========================================================================
-
-    #[tokio::test]
-    async fn test_execute_secret_sharing_validation_fails() {
-        println!("\n=== test_execute_secret_sharing_validation_fails ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let mut request = create_test_request(&crypto);
-        request.secret = vec![]; // Invalid
-        println!("  Testing with empty secret (should fail validation)");
-
-        let result = service.execute_secret_sharing(request).await;
-        println!("  Result: {:?}", result);
-        assert!(matches!(result, Err(WorkflowError::ValidationError(_))));
-        println!("  [PASS] Validation correctly failed for empty secret");
-    }
-
-    #[test]
-    fn test_get_secret_status_not_implemented() {
-        println!("\n=== test_get_secret_status_not_implemented ===");
-        let service = create_test_service();
-        let secret_id = SecretId::generate();
-        println!("  Checking status for secret_id: {}", secret_id);
-
-        let result = service.get_secret_status(&secret_id);
-        println!("  Result: {:?}", result);
-        assert!(matches!(result, Err(WorkflowError::ResourceNotFound(_))));
-
-        if let Err(WorkflowError::ResourceNotFound(msg)) = result {
-            assert!(msg.contains("not found") || msg.contains("not yet implemented"));
-            println!(
-                "  [PASS] ResourceNotFound returned (storage not implemented): {}",
-                msg
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn test_execute_with_different_threshold_total_combinations() {
-        println!("\n=== test_execute_with_different_threshold_total_combinations ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-
-        // Test various valid combinations
-        let combinations = [(2, 3), (3, 5), (5, 10), (2, 10)];
-
-        for (threshold, total) in combinations {
-            println!("  Testing (k={}, n={})...", threshold, total);
-            let mut request = create_test_request(&crypto);
-            request.threshold = threshold;
-            request.total_shares = total;
-
-            let result = service.execute_secret_sharing(request).await;
-            assert!(
-                result.is_ok(),
-                "Failed for threshold={threshold}, total={total}"
-            );
-
-            let result = result.unwrap();
-            println!(
-                "    kfrag_count={}, share_tx_ids.len()={}",
-                result.kfrag_count,
-                result.share_tx_ids.len()
-            );
-            assert_eq!(result.kfrag_count, total);
-            assert_eq!(result.share_tx_ids.len(), total as usize);
-        }
-        println!("  [PASS] All threshold/total combinations work correctly");
-    }
-
-    #[tokio::test]
-    async fn test_result_secret_id_is_unique() {
-        println!("\n=== test_result_secret_id_is_unique ===");
-        let service = create_test_service();
-        let crypto = CoreCryptoServiceImpl::new();
-
-        let request1 = create_test_request(&crypto);
-        let request2 = create_test_request(&crypto);
-
-        println!("  Executing two workflows to verify unique IDs...");
-        let result1 = service.execute_secret_sharing(request1).await.unwrap();
-        let result2 = service.execute_secret_sharing(request2).await.unwrap();
-
-        println!("  secret_id_1: {}", result1.secret_id);
-        println!("  secret_id_2: {}", result2.secret_id);
-
-        // Each execution should generate a unique secret ID
-        assert_ne!(result1.secret_id, result2.secret_id);
-        println!("  [PASS] Each execution generates unique secret_id");
-    }
-
-    // ========================================================================
-    // Partial Batch Failure Tests
-    // ========================================================================
-
-    #[tokio::test]
-    async fn test_execute_secret_sharing_partial_batch_failure() {
-        println!("\n=== test_execute_secret_sharing_partial_batch_failure ===");
-        let (service, mock_storage) = create_mock_service();
-        let crypto = CoreCryptoServiceImpl::new();
-        let request = create_test_request(&crypto);
-
-        mock_storage.set_should_fail_batch_partial(true);
-        println!("  Executing with partial batch failure enabled...");
-
-        let result = service.execute_secret_sharing(request).await;
-        assert!(result.is_err());
-
-        let err = result.unwrap_err();
-        println!("  Error: {:?}", err);
-
-        match err {
-            WorkflowError::PartialStorageFailure {
-                ref capsule_tx_id,
-                ref successful_share_tx_ids,
-                ref failed_shares,
-                failed_count,
-                total_count,
-            } => {
-                assert!(!capsule_tx_id.is_empty(), "capsule_tx_id should be present");
-                assert!(
-                    !successful_share_tx_ids.is_empty(),
-                    "some shares should have succeeded"
-                );
-                assert!(!failed_shares.is_empty(), "some shares should have failed");
-                assert_eq!(failed_count, failed_shares.len());
-                assert_eq!(
-                    total_count,
-                    successful_share_tx_ids.len() + failed_shares.len()
-                );
-
-                for (index_str, error_msg) in failed_shares {
-                    assert!(
-                        !index_str.is_empty(),
-                        "failed share index should not be empty"
-                    );
-                    let _index: usize = index_str.parse().expect("index should be a valid number");
-                    assert!(
-                        !error_msg.is_empty(),
-                        "failed share error message should not be empty"
-                    );
-                }
-                println!("  capsule_tx_id: {}", capsule_tx_id);
-                println!(
-                    "  successful_share_tx_ids: {} items",
-                    successful_share_tx_ids.len()
-                );
-                println!("  failed_shares: {} items", failed_shares.len());
-                println!("  [PASS] PartialStorageFailure contains all expected info");
-            }
-            other => panic!("Expected PartialStorageFailure, got: {:?}", other),
-        }
-
-        assert!(!err.is_recoverable());
-        println!("  [PASS] PartialStorageFailure is not recoverable");
-    }
-}
