+++ client/tests/unit/usecase/core/mod.rs
@@ -1,0 +1,1 @@
+mod crypto;
