+++ client/tests/unit/usecase/mod.rs
@@ -1,0 +1,4 @@
+mod core;
+mod dto;
+mod error;
+mod workflow;
