+++ client/tests/unit/usecase/workflow/mod.rs
@@ -1,0 +1,3 @@
+mod container;
+mod secret_recovery_service;
+mod secret_sharing_service;
